feat: split typing from type guards
